[feat] thunk, redux: decompose to modules folder
