+++ public/src/modules/redux-thunk/thunk.js
@@ -1,0 +1,6 @@
+import { createThunkMiddleware } from './createThunkMiddleware'
+
+const thunk = createThunkMiddleware()
+thunk.withExtraArgument = createThunkMiddleware
+
+export default thunk
