+++ public/src/modules/redux/combineReducers.js
@@ -1,0 +1,9 @@
+export function combineReducers(reducersMap) {
+	return function combinationReducer(state = {}, action) {
+		const nextState = {}
+		Object.entries(reducersMap).forEach(([key, reducer]) => {
+			nextState[key] = reducer(state[key], action)
+		})
+		return nextState
+	}
+}
