+++ public/src/modules/redux/compose.js
@@ -1,0 +1,13 @@
+export function compose(...funcs) {
+	if (funcs.length === 0) {
+		return arg => arg
+	}
+	if (funcs.length === 1) {
+		return funcs[0]
+	}
+	return funcs.reduce(
+		(a, b) =>
+			(...args) =>
+				a(b(...args)),
+	)
+}
