+++ public/src/redux/reducers/index.js
@@ -1,11 +1,3 @@
-function combineReducers(reducersMap) {
-	return function combinationReducer(state = {}, action) {
-		const nextState = {}
-		Object.entries(reducersMap).forEach(([key, reducer]) => {
-			nextState[key] = reducer(state[key], action)
-		})
-		return nextState
-	}
-}
+import { combineReducers } from '../../modules/redux/combineReducers'
 
 export const reducer = combineReducers({})
